requirements.txt
Django~=5.1.4
djangorestframework~=3.15.2
Faker~=33.1.0
factory_boy~=3.3.1
PyJWT~=2.10.1
gunicorn
psycopg2-binary
django-cors-headers